+++ components/Chunk_Types/include/DetectionBinChunk.h
@@ -40,7 +40,7 @@ public:
      * @brief Get the size of object in bytes
      * @return Size of object in bytes
      */
-    unsigned GetSize() override;
+    u_int64_t GetSize() override;
 
     /**
      * @brief Fill a byte array the represents this object
@@ -116,7 +116,7 @@ private:
      * @brief Get the size of object in bytes
      * @return Size of object in bytes
      */
-    unsigned GetInternalSize();
+    u_int64_t GetInternalSize();
 };
 
 #endif
