+++ components/Chunk_Types/include/WAVChunk.h
@@ -133,7 +133,7 @@ public:
      * @brief Get the size of object in bytes
      * @return Size of object in bytes
      */
-    unsigned GetSize() override;
+    u_int64_t GetSize() override;
 
     /**
      * @brief Returns if the two classes are equal
@@ -152,7 +152,7 @@ private:
      * @brief Get the size of object in bytes
      * @return Size of object in bytes
      */
-    unsigned GetInternalSize();
+    u_int64_t GetInternalSize();
 };
