+++ components/Chunk_Types/source/GPSChunk.cpp
@@ -29,14 +29,14 @@ GPSChunk::GPSChunk(const GPSChunk &GPSChunk) : BaseChunk(GPSChunk)
     m_bIsWest = GPSChunk.m_bIsWest;
 }
 
-unsigned GPSChunk::GetSize()
+u_int64_t  GPSChunk::GetSize()
 {
     return GetInternalSize();
 }
 
-unsigned GPSChunk::GetInternalSize()
+u_int64_t  GPSChunk::GetInternalSize()
 {
-    unsigned uByteSize = 0;
+    u_int64_t  uByteSize = 0;
 
     // First check baseclass
     BaseChunk SelfBaseChunk = static_cast<BaseChunk &>(*this);
